Feature/bc 540 implement new default profile image (#98)
### Feature: User Profile Image Enhancements

This pull request addresses the Jira ticket [BC-540/566 ] & Jira ticket
[BC-605 ] aim to improve the user experience related to profile images.
It includes frontend and backend changes to implement default profile
images based on a user's initials and allows users to edit their profile
pictures seamlessly.

### Changes Made:

1. Default Profile Image Implementation:

- A default profile image will be generated based on the user's first
and last initials.
- The initials will be displayed on a color background determined by UX
specifications.
- Implemented across various sections: navbar, sidebar, profile page,
edit profile, image modal, chat, and group chat.
- Styled the default profile image according to the design provided in
Figma.

2. Profile Picture Editing Functionality:

- Enabled the "Edit My Profile Picture" functionality on the "My
Profile" page.
- Integrated a flow allowing users to edit their profile picture,
including options to browse and upload a file, crop and edit the file,
and delete the file.
- Implemented logic to bypass the browse and select photo step when
editing an existing photo. Users can now directly edit the selected
photo, including cropping and centering, without the need to browse for
a new file.

### Review Notes:

Another ticket would need to be created to handle the loading state when
a user is changing the profile to avoid a slight lag in rendering. I
have already begun creating the additional logic to handle that but
could not finish since there are some quirks to fix in order to work
smoothly and the AC where already met and didn't want to drag the ticket
for another day.

Also will need to maybe create another ticket for a situation where when
a user doesn't save the profile picture uploaded should they be able to
close the modal til they do or not? Plus Might need to reach out to UX
on a flow that handles that possibility how should it display
style-wise?

### Deployment Notes:

No specific deployment requirements for this feature. The changes will
be included in the next deployment cycle.

**_Please review the changes and provide feedback as necessary. Thank
you!_**