BC-609/Revamp-My-Availability (#100)
This PR address the AC items defined in
https://bootcamper.atlassian.net/browse/BC-609
* Centers availability component on My Availability tab of calendar
component
* Verifies that all days are selectable
* Verifies that all days of the week are deletable
* Verifies that a user can copy time frames to other days of the week
* Address the bug where multiple copy time frame modals appeared in
tandem - now only one may appear at a time and clicking outside of the
modal will close it

Note: While working on this ticket I noticed other bugs in the
functionality and flow of this component. I'll make a new Jira with new
ACs to track that followup work.