Hifi/bc 641 project presentation completion (#129)
### This PR accomplishes:

- Styles the 4 pages involved in the Project Completion flow to match
the Figma HiFi.
- Componentizes the Project Url Input and Participation Radio Inputs for
reusability and proper state management.
- Refactors form submission logic to only update redux throughout the
flow and only perform the database update on final for
confirmation/submission.
- Adds more optional props to Primary/SecondaryButton to expand their
utility.
- Adds the `trackCompletion` prop to Paginator so that it can be used as
a stepper or to simply show location.

### This PR does not accomplish:

- Styling ProjectDetails screen to match Figma HiFi.
- The screen is close but there's definitely changes to be made that are
beyond nits
- It felt like I was getting into the weeds on this so it doesn't have
the "Your team will/won't be presenting" toast and the button/modal to
edit participation.
- Since I componentized the radio input it will be pretty quick to
implement in a future ticket.

### Questions/Considerations

- I know we're approaching a more serious push for componentization so
just pointing out that the ProjectDetails/Presentation tab is the exact
same content as the ProjectCompletion/Presentation page (minus the
inputs/buttons).
- Based on the talk I had with Dan, Erick, and Hector, it seems like the
data stored in the `completedInfo` of a project can be slimmed down
quite a bit.
- currently `deployedUrl = {userID: 'www.project.com'}`, not huge but if
we're only collecting one url that cannot be changed (per HiFi notes)
then I propose we simply store the url in a string and leave the userID
out.
- `completedInfo` also has a `participatingMembers` prop that is an
array with entries containing full user info and their decision to
participate or not. It was confirmed that only one team member goes
through the ProjectCompletion flow so I propose we remove
`participatingMembers` and just use a prop like `presenting` that stores
a boolean. This is currently implemented alongside with
`participatingMembers` prop.
- Currently the pagination on the ProjectCompletion flow allows users to
click ahead of the step they were one. I propose removing nav
functionality from the pagination in this use-case so the user must use
the buttons.
- I don't know where we're at on it yet, but we've got a TODO regarding
the google survey that users can take at the end of the flow and I just
wanted to check in on that.

Edit: I'm not sure what I did to include the commits from my first
ticket on this, but these changes begin on commit
[9b89963](https://github.com/yeezick/bootcampr/pull/129/commits/9b899632da8b345d2d34ad10bb7b44ea7e40cb8b)

[Ticket BC-641](https://bootcamper.atlassian.net/browse/BC-641)