Bugfix/bc 733 improve timezone handling availability bugs (#218)
This PR addresses various bugs in the availability component:

- 'Everyday' on the copy times modal now renders as 'Every Day'
- The plus sign to add a new time slot persists on the final time slot
for a day (previously the time slots were consolidating on render, which
affected whether the plus sign was visible or not)
- End time can no longer be before start time
- Logic improved to consolidate and sort time slots
- Fixed the bug where the end time slot goes blank / undefined following
certain flows
- Disables the add time slot button and logic if the end time is 11:00
PM or later

NOTE: I think there is still an edge case where the second slot goes
undefined - I came across it once but haven't been able to reproduce
again. The majority of these cases have been resolved, but if a dev or
QA is able to reproduce this bug, please take note of the steps to
reproduce! If given the steps, I can address in another branch / PR